[OMEMO] Add fallback message
If a recipient has published device keys but the current client doesn't
support OMEMO, it might not display an error or give any hint that there
was such a message.

This is even worse when I recently messaged a PSI 1.5 user with OMEMO
keys (made by other clients), and messages were not only completely
suppressed but the client even has sent back message receipts.

Adding a fallback <body/> to the message addresses this and it's also
the way most other clients with OMEMO support handle this.

Signed-off-by: aszlig <aszlig@nix.build>